Project Finished: Ready For Review
In outstanding quick time (for my standards), the project has already been finished in about the span of a mere hour, including all of it's CSS and documentation!

What we were instructed to enter in was a YouTube video for "The Road Not Taken", two audio files each for "Fire and Ice" and "Devotion" that include both a .mp3 and .ogg files, with also a fallback .webm file for browser plugins and a paragraph to even cover that if all else fails.

These projects were among some of my favorite and were certainly easier than most others, and I already look forward to using the skills within for further, more practical use.